requirements.txt
Flask==3.1.0
flask-cors
langchain==0.3.12
langchain-openai
wikipedia
python-dotenv
